Fix Home Assistant auto discovery of zones
To properly support HAss discovery of zones attached to the panel,
supply the "name " and "device" within the post to the config topic.
Additionally, send this message with the retain flag, such that the zone
entity is not in "unknown" state when home assistant is restarted and
can not find the config topic.

Signed-off-by: Matt Redfearn <matt.redfearn@gmail.com>